pyproject.toml
[project]
name = "sentiment-fused"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "gdown>=5.2.0",
    "python-dotenv>=1.1.1",
    "moviepy>=2.2.1",
    "speechrecognition>=3.10.0",
    "streamlit-webrtc>=0.47.0",
    "opencv-python-headless>=4.8.0",
    "torch>=2.8.0",
    "pillow>=11.3.0",
]
